Added Glue Data Catalog functionality
